Fix the Cleanup default window timeout setting
Drop any attempt to carry the default window show timeout setting in a static variable inside OutputWindow.  Instead save it as a member variable in CleanupConfigPage, so the settings will be read once per config dialog, and once for every OutputWindow that needs to know it.
Fix the config dialog itself to set the checkbox and spin-box based on the settings in the Cleanup itself, and the OutputWindow default as needed, and save them correctly.