One or two refresh crashes and other messes
Refreshes can throw exceptions if the top of the tree is no longer accessible, for example due to permissions or simply not being there any more, so catch the exceptions.
Refresh selected doesn't actually refresh the selected item, only its children.  This means that it may show obsolete data; will definitely show obsolete data if the permissions have changed or if it has been removed (eg. by a Cleanup).  If previous permissions issues have been fixed, a refresh of the selected item will fail to propagate the lack of errors up the tree and the tree may still show that there are permissions errors even though there aren't.
1) Process the refresh set to check that the items are still accessible and recurse up the tree until reaching a level where they are accessible.  Immediately just refresh the whole tree if the top level is reached.
2) Allow exceptions thrown while reading the whole tree during refresh selected to propagate to the main window and display the relevant dialog to show the error.
3) Delete the subtree on refresh selected instead if just clearing it, to update the directory properties.
Fix DirInfo::findNearestMountpoint() to stop at the top of the directory tree instead of <root>, so now it isn't necessary to explicitly fall back on the tree device.
Fix ignoreEmptyDirs() and moveIgnoredToAttic() to properly mark the (empty) children of empty directories as ignored and move empty directories to the attic in all cases.