Stote the row number directly on FileInfo
After the packing tweaks, there is space to squeeze in an int without increasing the structure size (on most modern architectures), so:
- add _rowNumber to FileInfo;
- remove the QMap of row numbers in DirSortInfo.
This speeds up model access to the row number, not dramatically, but enough to easily measure for some operations (try ctrl-5, ctrl-A).  Overall memory use should be down, but not enough to notice in most situations.