Fix column auto-size
Qt, by default, examines no more than 1,000 rows when auto-sizing a column in q QTreeView, and then sticks with that size.  So occasionally a visible row will have column contents (most commonly SizeCol) that don't fit in the automatically-chosen width.  When such rows are visible, the column will resize if something triggers a repaint.  This can be configured to more than 1,000 rows, but no value will ever fix the problem 100% and it starts to get a bit slow especially with all columns showing and set to auto-resize.  There was previously a hack in navigateUp() to try and get round this, but it didn't really fix the problem and was quite clunky.

This fix checks the width of the columns of all the rows visible in the viewport against the current column widths, for visible columns configured to auto-resize.  This happens every time the vertical scroll position changes (ie. whenever the rows visible in the viewport change).  If a column is found to not fit, a resize is triggered.  Not all columns are checked; for example the percent bar and permissions columns are always essentially the same width.  The check is made frequently, but only takes about a millisecond.

With all this monitoring, the 1,000-row lookahead could probably be reduced to make things snappier, but in practice it is not really a bottleneck even on a relatively slow machine.