Make dialog refreshes and syncs work better
In the stats and locate files dialog, make the initial read, refresh, and sync behaviour more consistent.  The item will be the current item, or its parent, or root as a fallback.
Make sync and refresh work across DirTree refreshes or complete new reads, including where there is no tree after the refresh.
Make the header url update to the current url (or fallback) after a sync or refresh, particularly after a new read.
Make the elided header label refresh when necessary on refreshes and syncs even if there is no actual resize event.