Maybe not better, try this
Calculate the required dialog width from the width of the longest row minus the width of the title, in terms of the width of a space character, and then pad with that number of non-breaking spaces.
This is still not precise because the title is rich text (<h3>), but it should be at least as wide as we calculated.  The extract effort of parsing the width of the rich text doesn't seem worthwhile.