Make DirTree:clear() safe
Instead of requiring that DirTree:clear() is only ever called from the DirTreeModel wrapper to safely reset the model, send signals before and after the clear so the model is always notified.  Now DirTree::clear() can be called directly and the case of a cache read automatically replacing the whole tree from DirReadJob is safe.