Now it is possible to make cache reads handle aborts more gracefully
When a cache read is aborted, or ends prematurely for some other reason, mark the top level of the cache directory tree as aborted.  Also remove the entire contents of the tree except for that top directory, since there is no practical way to know which directories have been fully read and which have not.