Cache file error handling
Don't allow a cache file to be written for a file without a directory parent (should probably not even be allowed in DirTree).  Disable the action and reject it in CacheWriter.
Don't allow any file entries in a cache file to be loaded into the tree until there is a valid top level to the tree.  Basically, all cache files should start with a directory, either as a replacement for a read in an existing tree, or as the top level of a new tree.