Prevent opening a single file as a directory tree
It is currently possible to open a single file as the toplevel of a tree.  This can be a symlink if the target does not exist, otherwise the symlink target is opened.  This is relatively meaningless, there is no percent data, it cannot be written to a cache, and the treemap has no parent to shade a cushion properly.
In startReading(), detect when the input is not a directory:
- a directory, which may be a target of a symlink that exists, will be canonicalised and read;
- for a non-directory that exists, including as the target of a symlink, canonicalise the parent directory;
- for a broken symlink, canonicalise the parent directory of the symlink (not the target);
- otherwise attempt to read the original (nonexistent)  input path, which will throw and produce an error message.

Fix so that the parent directory is opened