OutputWindow crash on shutdown
OutputWindow causes a segmentation fault when shutting down the program with Cleanup process still running.  The processes are deleted by their parent and then the destructor tries to delete them again.  Neither kills the processes first so it is very slow and messy anyway.
Create Cleanup processes with no parent so they can be managed cleanly inside OutputWindow.  Kill the processes first even in the destructor so they don't spawn lots of error messages.
Also slipped in making the background colour configurable.  The default is black.