QList/Vector
Change the QList queues in DirReadJobQueue into QVector so that they can be squeezed (without resorting to a custom method).  Qlist in Qt6 and QVector since about 5.7 don't release capacity when cleared, but QList doesn't have a squeeze() function until it is replaced by QVector in Qt6.  Squeeze both lists when they are cleared, the main queue in particular can get quite big.
Also modify the killQueue function that used QMutableListIterator.  There is no QMutableVectorIterator in Qt6.  It is also considerably simpler now.