Show an output window if Trash fails
With the ShowAfterTimeout policy, show the window immediately if there is an error instead of ignoring it and then deleting the window.  Now only ShowNever doesn'y show a window after an error.
This applies to Trash which always shows (or doesn't, if the Trash succeeds quickly) the window after a timeout (different from QDirStat).