Clean up Settings classes
A little more inheritance, a little more encapsulation, and some re-factoring.
The group prefix is no longer explicitly stored, but hard-coded for each of the derived classes.  It is also never supplied externally so the API is a little simpler.
There is also no overloaded beginGroup() function, which means it automatically benefits from the QAnyStringView improvements in Qt6.  The related beginListGroup() function now just takes an int and gets the prefix from the class.
There is just one migrate function, in the base class, which uses the fileName() and group prefix supplied by a derived class to do the right thing.
migrate() is now a little faster in the default (no migration) case, but still a bunch of work that is probably never going to help anyone.