Gile type statistics "rewrite"
The whole top-X codepath wasn't being reached since the NO_SUFFIX/NON_SUFFIX changes.  Rewrite as a simple prune of the "Other" category list.  Make topX configurable instead of a define.
Don't zero out the no-extension stats before moving cruft.
Change the hash maps to contain both counts and sums, and drop the separate non-suffix rule maps, so now there are just two maps.  Define structs to replace the pairs so that their members can be accessed by name instead of just first and second.  FileTypeStatsWindow doesn't have to have to do separate lookups for the sums while iterating through the counts, and it doesn't need to do special handling for the non-suffix rules.
Drop the special strings for identifying NO_SUFFIX and NON_SUFFIX.  They are just empty strings.  FileTypeStatsWindow can map empty suffix strings to the text it wants to display, which it was doing anyway from NO_SUFFIX and NON_SUFFIX.  All the special handling for identifying these two cases becomes isEmpty().
Tighten up the cruft rules slightly to avoid merging somewhat-meaningful extensions.  Make the "uncommon" check both in absolute and relative terms: "2" in a large tree could be extremely uncommon, while in a small tree it might mean it is quite common.  Treat "1" as uncommon in all cases.  Don't compare the count with the length, that's not a very meaningful check.  Make the "long extension" check 16 characters since some extensions like ".introspection" are quite long.  Keep the letters percentage check at 75% although it might be better at 50%.
Tweak the logging, hide the more verbose cases behind a define and add a summary line at the end of the stats collect.
The code is now dramatically smaller and considerably faster.
In a similar vein, add structs to replace the QPairs in MimeCategorizer (WildcardPair) and FileInfo (YearAndMonth).