Another round of string changes, plus bits and pieces - headers
Use _L1 to mostly replace QLatin1String literals.  Use u'' for Latin1 character literals.  This seems to be the Qt-preferred approach at the moment, with a view to deprecating QLatin1Char, although it triggers ambiguous overloads in some cases (eg. .arg()).
Add _L1 to some more string literals now that it is less cumbersome.  It is marginally better than char * in QStringBuilder, even though they are essentially the same thing, possibly because QLatin1String is concatenable as a QString rather than QByteArray.  Perhaps it is only better when merging with QStrings.  Anyway, the code is smaller, so _L1 it is.  In most other overloads, const char * seems (marginally) better.
Replace a few remaining uses of parentheses for member variable initialisation, except the one famous vector case.
Refactor DpkgManager calls to runCommand().  Change runCommand() in SysUtil to have a logError argument insterad of ignoreErrCode.  Adjust defaults and calls to match.