More PercentileStats encapsulation
Replace the wrappers in HistogramView.  Keep percentile() as a convenience.  Do sanity checks on all input indexes and percentiles.
Store the bucket list start and end percentiles in PercentileStats so they don't need to be re-validated every time, bucketWidth() can be an inline function in PercentileStats, and bucketStart() and bucketEnd() can also be in PercentileStats.
Now BucketsTableModel can access the stats directly and doesn't need to know the HistogramView.  HistogramBar also accesses the stats directly in the constructor.