Impose file count and size limits - headers
Create typedefs ("using") for file counts and dir sizes as well as for file sizes, and corresponding "max" defines.  Keep the overall limit and per-dir limit as 32-bit signed int.  Two billion files could possibly be exceeded, but it would be impractical to run against filesystems bigger than that.  Keep the file size limit as a 64-bit signed integer, again could conceivably be reached but probably impractical to manage anything bigger.
Add checks in childAdded() that those limits are not breached.  Raise exceptions, nothing fancier for unlikely errors.
Clean and (and fix) the summary counting, especially of the un/ignored items which were double-counted at times, just wrong some of the time, and pointlessly updated when _summaryDirty=true.
Analyze narrowing conversions throughout the codebase, tweak some data types, and add a few casts.  Most of the narrowing conversions are either intentional or unavoidable  and not a problem (eg. long long to double in TreemapTile).  qsizetype to int in Qt6 is also largely ignored since those containers can't be larger than an int in this program.  Consider usisng something like size_t for loop counters on containers.