DirInfo summary counts
Fix recalc() to not count dot entries in _totalItems.  This makes it consistent with the way childAdded works (it doesn't see dot entries, so doesn't count them) and is one less reason for forcing massive recalcs.
Fix checkIgnored to mark _summaryDirty when the status changes.
Remove the DirTree recalc() function (or at least calls to it) that recalculates the whole tree because the counts should be in sync now, or at least marked dirty.
When deleting an empty dot entry, just decrement _directChildrenCount instead of re-counting everything.
Don't count un/ignored items when the summaries are dirty, they might be wrong anyway.