TreemapTile work again
Make FileInfoIterator slightly simpler.  Or at least the increment operator is simple because it is explicitly initialised instead of having to check every time, and the flags are gone.
The treemap build thread currently calls FileInfo functions that are not const (DirInfo::_summaryDirty checks and recalcs).  This doesn't cause problems because the subtree is clean when the build thread starts and it is killed before anybody can change anything.  Still, formalize this by only calling const functions in FileInfoBySizeIterator and the treemap build thread just in case anyone finds a way to dirty the tree while the thread is running.