Make locate windows appear promptly
LocateFilesWindow and LocateFileTypeWindow select the first item in the list because this doesn't consistently happen otherwise.  This slows down the display of the window by a surprising amount; an action that takes less than a millisecond at other times takes many tens of milliseconds at this time, presumably in difficulty because the list hasn't yet been rendered.
So put the selectFirstItem call in a QTimer.  The window opens instantly, although it then takes 50ms for the first line to get highlighted.