More on dialogs, more on PercentileStats
Re-factor the discover dialogs some more.  Tweak the delay on selecting the first line.
Make the PercentileStats data value type qint64 since the data that is stored are all integers.  The sums are also integers but the _percentiles boundaries list are long double.  Pay close attention to rounding and narrowing of this and related variables.
Change the TreeWalker class model to restrict _overflow to the class that uses it, make overflow() virtual and hard-coded to false except for one class, make prepare() virtual and a no-op for the base class (no longer _overflow in the base class) and override functions don't need to call the base implementation, and tweak the percentile stats analysis and rounding.