DotEntry.cpp is back!
Fix locate(), especially for searching within attics.
Put the attic- and dot-entry-specific code into virtual function overrides instead of inside if statements in FileInfo.  PkgInfo already has an override, but it is trivial.
Drop the findPseudoDirs flag.  If the url ends in a pseudo-dir then it will be searched for, otherwise not.