Treemap refactoring
Move the static local variables from renderCushion into TreemapView.  Re-calculate them every time the settings are changed (as opposed to never), not that the only relevant setting has any ui at the moment.  _lightX, _lightY, and _lightZ are now the coefficients adjusted for the ambient light.
Store ambientIntensity (a 0..1 double) instead of ambientLight (a 0..255 integer) since that is what TreemapTile uses.