Another attic refresh edge case
When an ignored item is refreshed and an unignored item is in the new subtree, it still appears (not greyed out) in the ignored tree.  Move ignored items out of the attic before they are refreshed.  They will usually get moved back after the refresh, but any unignored items will then be outside the attic.
The fix for the previous issue can now be a little simpler: just mark any attic dirty when it gets a directory moved into it; childAdded can't handle errSubDirs, can't handle any populated directory insert really.