Attic refresh cleanup
Improve the notifications to the model for unatticOne().  The only insert notification is "all children" so tell the model that all the children are deleted (then possibly deleting an empty attic and re-inserting the child from the attic) before telling it they're all back, and then possibly deleting one of them again!  This is all to avoid disturbing the expanded indexes as much as is possible when refreshing chunks of subtree.
Make it clear that childAdded() does not understand or deal with additions that have children, and make sure that everyone adding subtrees or deleting items marks the relevant items as dirty and drops the sortInfo.
Re-factor DirTree::stat() to a flatter and simpler layout.