More config changes
Save the config dialog window geometry and position to Settings.
Add context menus for Cleanup and Exclude rules, Mime Category already had one.  Standardize the actions and menu in ListEditor to work with the tool buttons.
Rearrange the Cleanup hotkey and icon fields now that there is a little more vertical room, so they are less crowded horizontally.
Use QDialogButtonBox for better platform integration and consistency (not with every window, more to follow!).