Make Refresher more resilient
Refresher::refresh() blindly assumes that the first item in its list is still valid (although it checks that tree() is not null!).  Add a magic number check since it is very easy for something to have been refreshed or completely replaced while a Cleanup was running.  I'm going to close off some of the more obvious ways to to do this, but it is best that Refresher doesn't crash if something sneaks through.