Configurable hotkeys
Look up each action in Settings: if there is an entry, load the value as a QKeySequence.  If it is valid (including empty), load that as the new shortcut on the action.
If there is no valid settings entry, save the shortcut from the action to the settings.
ActionManager processes all the MainWindow actions. FilesystemsWindow has two actions and FileTypeStatsWindow has two.