Sort out Percentiles data types
The boundaries of percentiles in PercentileStats are calculated and stored internally as long doubles.  The size stats dialog only needs them as integers (TreeWalker also only needs integers, but sometimes rounds up and sometimes down).  By a fluke, the PercentileStats::percentileBoundary() truncates the long double and returns an integer, and everything works.
Instead, return the internal floating point value but also provide a wrapper that formally calls std::floor and returns an integer.
The main need for these long doubles is for buckets.  Tweak the fillBuckets() loops to be slightly more efficient.
Change the algorithm for calculating the percentile boundaries to a "C=1" interpolating algorithm that doesn't require the min and max limits to be handled as special cases and provides more statistically valid results between values especially for small datasets (although they are mostly used as integers anyway).