File size stats window geometry
Drop the width constraint in calcGeometry().  The width is already constrained in the .ui file and the additional constraint is not helpful.  The lower height limit is set to approximately match the height of the overflow panel so that they scale together.
Drop the height constraint in the .ui file.  No real reason to restrict the user if they want a tiny histogram.  The width remains constrained to the width of the options panel.
Change the upper height limit in calcGeometry() to the width; allowing the aspect ratio to become taller is not helpful.