Fix refresh update of the histogram
The histogram does not rebuild if it is not visible.  This is to prevent undefined behaviour (ie. a crash) if it tries to rebuild before the window geometry is known.
Change this to be if the window is not visible.  The histogram should update on a refresh even if it is not currently the visible tab.