Include symlinks in the stats reports
Simplify the is...() functions in FilerInfo.h and FormatUtil.cpp.
In MimeCategorizer, make the caller of the protected category() function responsible for initializing suffix_ret.  The only callers are the two public category() functions and only one of them uses suffix_ret.  Also make category(FileInfo *, QString *) detect symlinks.
Make FileAgeStats, FileSizeStats, and FileTypeStats collect for symlinks as well as regular files.  This increases counts for the relevant months or years for FileAgeStats, adds a Symlinks category entry near the bottom of the list for FileTypeStats, and bloats out one or two of the lowest buckets for FileSizeStats.  Other non-dir object types do not have sensible sizes, ages, or mime types.