y-axis labels in histogram
Add labels for the start and end of the y-axis in the file size histogram.  The start label is always "0", but the end label is useful, indicating the number of files level of the histogram.  It actually shows the value of the largest bucket.
Re-factor text handling in the histogram a little.
Provide more getters in PercentileStats, for the values of the quartiles and the min and max percentiles.  Make the getters constexpr static so they can be used in switch statements.
Avoid a (highly unlikely) integer overflow by testing the result as a float before trying to use it as an integer.