Histogram
Size the width of the overflow panel based on the actual width of the headline to allow for theming and font size variation.  If any of the other lines get wider than this (eg. very long cutoff numbers) then they will wrap.
Allow _minHeight to decrease, to allow for (un-)wrapping of lines in the overflow panel.
Increase the size of the pies slightly and flatten their appearance by removing the outline pen.
Center the text and pies in the overflow panel.
Change all the brush and pen variables to getters.  This allows for some to be const references and avoids carrying them permanently for one or two uses in each build.
Fix the cutoff values in the overflow panel to match the actual cutoff files and sizes rather than "nominal" values based on average percentile widths.
Make the auto log scale less aggressive: make the reference bucket count a minimum of one since zero will always yield a log scale; make the comparison 20x instead of 10x.  50x would also be just about OK, but is a bit more jarring if the scale switches between linear and log when tweaking the percentile ranges.