Histogram resizing
Finally get resizing right: in calcGeometry(), up-scale the width when the height is forced to _minHeight, so that after scaling in fitToViewport(), the content fills the viewport.  The histogram now more or less fills the viewport at any scaling; the overflow panel contents are scaled down from their natural size, but never up; the overflow panel does get additional padding at the bottom when the histogram height is increased; the histogram itself now expands to fill the remaining space; both the histogram and overflow panel are scaled down, preserving aspect ratio, when the overflow panel contents cannot fit in the viewport height.  The histogram can now become extremely wide compared to its height, but if that's what users decide to do ...
There is now a hard-coded minimum height for the histogram (ie. an initialisation value for _minHeight) so stop the histogram bars becoming inverted when there is no overflow panel to impose a sensible height.
Simplify the _minHeight setting at the end of addOverflowPanel().  The histogram is rebuilt whenever _minHeight is increased *or* decreased, although it should be very rare that any of the overflow panel lines wrap and change its height.
Make addPie() and addText() named lambdas inside addOverflowPanel().  They are specific to the overflow panel and should not be used elsewhere.