Formalize histogram initialisation
Show the window before populating the stats.  This triggers two resize events that would otherwise cause extra rebuilds of the histogram.  HistogramView is now guaranteed a fully-realized window when it is initialised.
Move autoStartEndPercentiles() into FileSizeStatsWindow so it has complete control over them.
Block the slider signals when setting the start and end percentiles during initialisation.  Otherwise these might (or might not if they happen to set the default values) send the star and end percentiles to HistogramView and call rebuilds.
Explicitly set the start and end percentiles to whatever autoStartEndPercentiles has come up with, and rebuild (once!).
No need now to check if the values have changed in startValueChanged and endValueChanged because they really have changed or the signal wouldn't be triggered.