Histogram context menu
Bit of a beast since there were no actions before, just widgets connected to functions.  Now the percentile markers (the median and quartiles are still fixed, I think they are near-essential for interpreting the histogram), start and end percentiles, and log scaling can be controlled from the menu and using configurable hotkeys.
The percentile increment and decrement actions are mostly on the menu just to make the hotkeys more discoverable and configurable.
Possibly the main reason for the context menu is to make it possible to manually control the log/linear scaling on the y-axis, very informative and worth trying.
The hotkeys are in the config file, everything else is per-dialog or per-stats.