And now that everything is re-structured ...
Dramatically simplify the triggering of start and end percentiles.  Simply call the setPercentileRange() function which passes both startPercentile and endPercentile values from the sliders to HistogramView, which automatically rebuilds.  Do the same at initialisation, after autoPercentileRange() which was blocked from sending signals.
Hand off validation of the start and end percentiles to PercentileStats which has to do the same validation for fillBuckets().