Event handling for highlight colouring
Add PaletteChanged to the list of events that trigger the list widget background to be redrawn.  The show and resize events seem to react to most theme changes, but I don't think it is guaranteed.
Also change the highlighting from a custom colour to the Mid role colour and change the text colour to HighlightedText to match.  This means that the current row will usually be blue; this is consistent with other config pages and hopefully doesn't clash too badly with the colour previews.