Palette changes in the file size dialog
Respond to palette change events in the histogram.  The histogram was responding because most theme changes trigger a resize event, but I don't think that is reliable.  The percentile table decile shading wasn't reponding at all.
Add changeEvent() overrides in FileSizeStatsWindow and HistogramView.
Change the HistogramView resize event handler to rebuild in a timer; rebuilding inside the event handler is not safe in some cases and the Qt code will crash if the children list is changed while it is recursively showing them.