Subscripts, tooltips, and more
Use subscripts for "Pn"-type text in the histogram.  Use "large" subscript text because standard subscripts in Qt can be tiny, especially with small fonts.  This is done with setHtml() on the graphics items.
Use similar subscripts in the percentile table vertical header, and also use subscripts in the horizontal header.  This requires sub-classing QHeaderView and painting each section directly.  Just use one class for both the horizontal and vertical headers, they are almost identical and QHeaderView inherently supports both.
Also display the table itself through a model.  This makes it simple to blank the default header text, and also to add tooltips.  The model itself is very straightforward, although the mapping of model rows to percentiles in the filtered view is tricky to support the "margins".  Make the first column into a vertical header ... well ... because it is.  Make the table a QTableView obviously.
Add tooltips to size columns where a rounded size is displayed (ie. >= 1,000 bytes) in both the percentile and buckets tables.  Add tooltips to the column headers of the percentile and buckets tables.  This should go a long way towards making these comprehensible without looking at the documentation.
Rename BucketsTableModel to FileSizeStatsModels and put all three models in the same h/cpp files.