Reposition the HistogramView options button
Place the button below the options panel so that it doesn't move about as the label opens and closes.
Also make it clickable instead of swapping the arrows and having one button inside the panel and one outside.  The visual results are more theme-dependant now, but it reduces the tangle of signals and button swaps.