Buckets boundaries
Straighten out the reporting of bucket start sizes so they are more consistently "inclusive".  Move the +1 calculation into PercentileStats so the histogram bars also get it, although it is even less visible there because only the rounded size is shown in the tooltip.
Handle special cases of P0 (P0 equals the smallest file, but that file is considered to be in the next percentile - making P0 smaller would mean it was sometimes negative and would not be calculated naturally) and empty buckets with zero width (when many files have exactly the same size).  Do not increment the start size for these: P0 is already inclusive for buckets that start at P0; when all the files (or none!) in a bucket have the same size, the start and end sizes are the same.
Beef up the validation of bucket indexes (not that they can ever be wrong!) and drop the rather generic CHECK_INDEX exception macro (and rename CHECK_INDEX_MSG to CHECK_INDEX) so there is always exception text specific to the type of invalid index.
Remove ResizeToContents mode from the vertical header. Having it set on both the horizontal and vertical headers can lead to resizing glitches.  Thanks for that Qt.  The vertical header sections are largely the correct height for the theme/style and do not change.  Make extra effort to center the labels vertically and align them horizontally within margins.
Fix scene switching in HistogramView.  Use deleteLater().