Log bucket widths
Re-factor PercentileStats to store bucket start values in a list instead of storing a start and end value to calculate them on the fly.
Use this mechanism to provide log-scaled bucket widths: buckets which increase in width exponentially.  There are the same number of buckets for a given set of stats linearly of log-scaled, but log-scaled has very small buckets at smaller file sizes and much wider buckets at larger file sizes.