Percentile table counts and sums
Make the count and sum columns show the values from the previous *visible* percentile to this one.  Showing values for a single percentile when the table is filtered and only every 5th percentile is visible is confusing and unhelpful.
Drop the whole "margins" idea from the table; also confusing and unhelpful, even more so when most of the rows in the table are showing counts and sums for 5 rows.