Expand buckets table
Add a bucket width column.  Not usually very informative for fixed-width buckets, but interesting with log bucket widths.  Keep it visible for both because there are some edge cases where it can clarify what is going on.  Use decimals for small bucket widths; widths less than 1 byte in particular could be confusing otherwise.
Keep the value column last since that is key information and it would be harder to spot if buried in the middle somewhere.
Add tooltips to the start and end columns to show the exact (to two decimal places) bucket boundaries when the bucket width is very small.  This can clarify why some buckets with apparently the same start and end have different contents.  Given the additional clarity, drop the special handling and allow the end size to be smaller than the start size: the start size is rounded up and the end size effectively rounded down to give inclusive bucket size limits, but if the bucket doesn't spam an integer size then the end size rounds down to one less than the rounded-up start.  Force the end to match the start for the very special zero-width bucket, the single bucket when all files are the same size.
Add a tweak to fillBuckets() so that with a start size of zero and log bucket widths, the last bucket doesn't end prematurely (because the bucket widths increasing by a factor effectively doesn't start until the second bucket because the first is not really a log).