Improved management of long labels in dialogs
The elideLabel logic makes assumptions about the width of the fixed portion of the label, including padding, when calculating the length to elide to.  This can fail with larger-than-expected fonts and the label reverts to glowing up the width of the dialog and refusing to resize down.
Take the fixed portion of labels out of the ui file and work with the entire string in elideLabel.  Now only the padding needs to be allowed for and that is in pixels.  It could still fail if someone messes with css and pads out too much, but that's something they'll have to accept.
Make some minor styling changes to panel frames, lines, and some tree header titles.