Fix OutputWindow hiding
Before Qt 6.3, closing a dialog via reject() actually just hides it.  This happens, for example, when the Escape key is pressed.  There is no close event and OutputWindow doesn't know the window is gone.  If there is more output, the window reappears.  If there is no more output, the user has lost any way to access the window, properly close it, or even to know it still exists.
So override the hide event instead, which is also generated when the window is actually closed.  Ignore "spontaneous" hide events, which happen in cases such as the window being iconified or being placed in a different workspace to the current one.  This still works in Qt 6.3 because a hide event is still generated.