Fix output window hotkeys
The hotkeys for the zoom buttons don't work because they aren't added to the window (although they are children of the window).  Use the same mechanism as in FileSizeStatsWindow, now as a static function in ActionManager, to check the hotkey settings and then add them to the window.  So now they're configurable although they are fairly standardised.
Make the actions trigger the buttons instead of the other way round.
Remove the kill action; it isn't used and having a hotkey beyond the alt-k accelerator is unnecessary.
Remove auto-default from the kill button, slightly dangerous.