isLower()/isUpper()
isLower() and isUpper() don't exist in Qt 5.11 and earlier.  Create simple inline replacements/wrappers in FormatUtil.h.
The replacements are used for all versions earlier than 5.14 because isLower() and isUpper() when first introduced were defined to only be true when the string wasn't empty and all characters were respectively lower or upper case letters.  The later versions and the replacement are defined to, for example only return false for isLower() when a character is considered to be (Unicode) upper case.
Tweak the core MimeCategorizer case-insensitive hash test to give overall slightly improved performance.