Enhance FileAgeStatsWindow and FileTypeStatsWindow
Add percent bars to FileTypeStatsWindow, using the same settings as FileAgeStatsWindow.  Move common code into PercentBar for want of a better place.
Expand the number of years that have month children in FileAgeStatsWindow.  Dramatically simplify FileAgeStats to get rid of all the complicated initialisation of static arrays and post-processing.  Just throw every collected file into a month hash and a year hash (the year hash isn't strictly necessary, but it costs little and simplifies the work for the dialog), with the value being a simple count/size pair.  Remove the year and month fields from the hash value struct; if you have the struct then you know the year/month.  Now the window can choose exactly which years to show month children; for now it is five, but it probably needs to be a little more intelligent or be configurable.