Short-circuit NTFS checking slightly
If there are no NTFS volumes then there is no point making slightly flaky block-size checks.  Add MountPoints::_hasNtfs, populate it when the NTFS devices are found, and check it from LocalDirReadJob.
Checking for BTRFS (once the map is populated) is hardly onerous, so just do it instead of trying to get clever with a lazy getter.