More robust crossing mountpoint checking
DirReadJob and DirTree both check to see if a child is a mountpoint that causes filesystem crossing; and they use different algorithms!  This causes subtle cases such as refreshing a mountpoint inside a tree read from a cache file to be flagged incorrectly.
Move the check into a single (static) function in DirTree.  Check if the devices match, that's not crossing even if they're both zero.  Then check if the child is actually a mountpoint, obviously not crossing if it isn't.  Then check if the device of the nearest mountpoint of the parent matches that of the child.